move write_fragalysis_output to the fauxalysis_from_docking.py script
